backend/requirements.txt
Flask
Flask-CORS
google-api-python-client
google-auth
google-generativeai
python-dotenv
requests
